# server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSave",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/sync/transactions/inbound": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Start Inbound",
        "operationId": "start_inbound_api_v1_sync_transactions_inbound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/sessions/{session_id}/manifest": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Post Manifest",
        "operationId": "post_manifest_api_v1_sync_sessions__session_id__manifest_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/sessions/{session_id}/window": {
      "put": {
        "tags": [
          "sync"
        ],
        "summary": "Upload Window",
        "operationId": "upload_window_api_v1_sync_sessions__session_id__window_put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/sessions/{session_id}/commit": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Commit Upload",
        "operationId": "commit_upload_api_v1_sync_sessions__session_id__commit_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/sessions/{session_id}/resume": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Session Resume",
        "operationId": "session_resume_api_v1_sync_sessions__session_id__resume_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/queue": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Poll Queue",
        "operationId": "poll_queue_api_v1_sync_queue_get",
        "parameters": [
          {
            "name": "sync_generation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Sync Generation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/device-config": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Device Config",
        "description": "Bootstrap endpoint (X-Device-ID only \u2014 no Bearer token required yet).\n\n1. Always: upsert reported known_profiles into device_known_profiles cache.\n2. If config_pending=1 AND devices.last_seen is within 15 minutes:\n   return {\"device_token\": \"<token>\"} and clear the flag.\n3. Otherwise: return {}.",
        "operationId": "device_config_api_v1_sync_device_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConfigBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceConfigBody": {
        "properties": {
          "known_profiles": {
            "items": {
              "$ref": "#/components/schemas/KnownProfile"
            },
            "type": "array",
            "title": "Known Profiles",
            "default": []
          }
        },
        "type": "object",
        "title": "DeviceConfigBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InboundBody": {
        "properties": {
          "title_id": {
            "type": "string",
            "title": "Title Id"
          },
          "total_size_bytes": {
            "type": "integer",
            "title": "Total Size Bytes"
          },
          "hardware_type": {
            "type": "string",
            "title": "Hardware Type",
            "default": ""
          },
          "parent_sequence_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Sequence Num"
          },
          "preservation": {
            "type": "boolean",
            "title": "Preservation",
            "default": false
          },
          "user_key": {
            "type": "string",
            "title": "User Key",
            "default": ""
          },
          "user_display": {
            "type": "string",
            "title": "User Display",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title_id",
          "total_size_bytes"
        ],
        "title": "InboundBody"
      },
      "KnownProfile": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "profile_name": {
            "type": "string",
            "title": "Profile Name",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "KnownProfile"
      },
...
      "ManifestBody": {
        "properties": {
          "checkpoint_size": {
            "type": "integer",
            "title": "Checkpoint Size"
          },
          "checkpoint_ledger": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Checkpoint Ledger"
          }
        },
        "type": "object",
        "required": [
          "checkpoint_size",
          "checkpoint_ledger"
        ],
        "title": "ManifestBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}